requirements.txt
chromadb>=1.0.7
numpy>=2.0.0
openai>=1.40.0
pypdf>=4.2.0
python-docx>=1.1.2
PyYAML>=6.0.1
python-dotenv>=1.0.1
requests>=2.32.0
rich>=13.7.1
tenacity>=8.2.3
tqdm>=4.66.4

fastapi>=0.112.0
uvicorn>=0.30.0

pytest>=8.2.0
